chore: migrated full handling strategy to raw paper access
