+++ api/src/main/java/net/paulem/krimson/common/Versioning.java
@@ -1,4 +1,4 @@
-package net.paulem.krimson.common.compat;
+package net.paulem.krimson.common;
 
 import org.bukkit.Bukkit;
