+++ api/src/main/java/net/paulem/krimson/paper/compat/serialize/PaperItemSerializer.java
@@ -1,45 +1,40 @@
 package net.paulem.krimson.paper.compat.serialize;
 
+import net.paulem.krimson.common.KrimsonPlugin;
+import net.paulem.krimson.paper.compat.stream.input.PaperInputStream;
+import net.paulem.krimson.paper.compat.stream.output.PaperOutputStream;
 import org.bukkit.inventory.ItemStack;
-import net.paulem.krimson.common.compat.serialize.ItemSerializerHandler;
-import net.paulem.krimson.common.compat.stream.input.InputStreamHandler;
-import net.paulem.krimson.common.compat.stream.output.OutputStreamHandler;
 import org.bukkit.plugin.java.JavaPlugin;
 
-public class PaperItemSerializer extends ItemSerializerHandler {
-    public PaperItemSerializer(JavaPlugin plugin) {
-        if (!messageSent) {
-            plugin.getLogger().info("Hooray! Using Paper's ItemStack (de)serialization, which gives you much better performance than Bukkit's!");
-            messageSent = true;
-        }
+public class PaperItemSerializer {
+    public static final PaperItemSerializer INSTANCE = new PaperItemSerializer(KrimsonPlugin.getInstance());
+
+    private PaperItemSerializer(JavaPlugin plugin) {
+        plugin.getLogger().info("Hooray! Using Paper's ItemStack (de)serialization, which gives you much better performance than Bukkit's!");
     }
 
-    @Override
-    public void serializeAndWrite(ItemStack stack, OutputStreamHandler<?> outputStream) throws Exception {
+    public void serializeAndWrite(ItemStack stack, PaperOutputStream outputStream) throws Exception {
         byte[] serialized = stack.serializeAsBytes();
 
         outputStream.writeInt(serialized.length);
         outputStream.write(serialized);
     }
 
-    @Override
-    public void serializeAndWrite(ItemStack[] stacks, OutputStreamHandler<?> outputStream) throws Exception {
+    public void serializeAndWrite(ItemStack[] stacks, PaperOutputStream outputStream) throws Exception {
         byte[] serialized = ItemStack.serializeItemsAsBytes(stacks);
 
         outputStream.writeInt(serialized.length);
         outputStream.write(serialized);
     }
 
-    @Override
-    public ItemStack readAndDeserialize(InputStreamHandler<?> inputStream, int length) throws Exception {
+    public ItemStack readAndDeserialize(PaperInputStream inputStream, int length) throws Exception {
         byte[] itemBytes = new byte[length];
         inputStream.read(itemBytes);
 
         return ItemStack.deserializeBytes(itemBytes);
     }
 
-    @Override
-    public ItemStack[] readAndDeserializeList(InputStreamHandler<?> inputStream, int length) throws Exception {
+    public ItemStack[] readAndDeserializeList(PaperInputStream inputStream, int length) throws Exception {
         byte[] itemBytes = new byte[length];
         inputStream.read(itemBytes);
