+++ api/src/main/java/net/paulem/krimson/paper/compat/stream/input/PaperInputStream.java
@@ -1,42 +1,38 @@
 package net.paulem.krimson.paper.compat.stream.input;
 
-import net.paulem.krimson.common.compat.stream.input.InputStreamHandler;
-
 import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
 import java.io.IOException;
 
-public class PaperInputStream extends InputStreamHandler<DataInputStream> {
+public class PaperInputStream {
+    protected final ByteArrayInputStream inputStream;
+    protected final DataInputStream dataInput;
+
     public PaperInputStream(ByteArrayInputStream inputStream) {
-        super(inputStream);
+        this.inputStream = inputStream;
+        this.dataInput = create();
     }
 
-    @Override
     public DataInputStream create() {
         return new DataInputStream(inputStream);
     }
 
-    @Override
     public Object readObject() {
         throw new RuntimeException("Paper's input stream handler doesn't support reading objects!");
     }
 
-    @Override
     public int readInt() throws IOException {
         return dataInput.readInt();
     }
 
-    @Override
     public String readUTF() throws IOException {
         return dataInput.readUTF();
     }
 
-    @Override
     public int read(byte[] bytes) throws IOException {
         return dataInput.read(bytes);
     }
 
-    @Override
     public void close() {
 
     }
