+++ settings.gradle.kts
@@ -11,5 +11,4 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
 }
 
-include("common")
 include("api")
